Extract config loading and file discovery into carina-core::config_loader
Move load_configuration(), LoadedConfig, get_base_dir(),
find_crn_files_in_dir(), and find_crn_files_recursive() from
carina-cli main.rs into a new carina-core::config_loader module.

Part of #328 (Phase 2: Extract orchestration logic).
